raw/mvp_site_all/requirements.txt
# Core dependencies (cache key: 2026-01-27-invalidate)
Flask==3.0.0
gunicorn==23.0.0
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
google-api-core>=2.19.0
google-auth>=2.35.0
cachetools>=5.3.0
google-genai
protobuf>=4.25.3,<5.0.0
grpcio>=1.62.0,<2.0.0
grpcio-status>=1.62.0,<2.0.0
# Updated for compatibility between protobuf, grpcio, and firestore
flask-compress>=1.15
Flask-Cors==6.0.0
Flask-Limiter>=3.5.0
fpdf2
python-docx
google-cloud-secret-manager
selenium
pydantic
jsonschema>=4.21.0
requests>=2.31.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
psutil>=5.9.0
PyJWT>=2.8.0
ruff>=0.6.0
isort>=5.13.0
mypy>=1.8.0
bandit>=1.7.0
types-flask>=1.1.6
types-requests>=2.31.0
mcp>=1.0.0
pytest>=6.0.0
pytest-timeout>=2.4.0
pytest-testmon>=2.2.0,<3.0.0
pytest-cov>=4.0.0
fastembed==0.7.4
numpy>=1.24.0,<2.2.0
